refractor: move Session class to dedicate module
